+++ src/imio/dashboard/tests/test_vocabulary.py
@@ -1,11 +1,14 @@
 # -*- coding: utf-8 -*-
 from zope.component import queryUtility
+from zope.interface import alsoProvides
 from zope.schema.interfaces import IVocabularyFactory
 from Products.CMFCore.utils import getToolByName
 from plone.app.testing import login
 from plone import api
 from imio.dashboard.testing import IntegrationTestCase
 from collective.behavior.talcondition.interfaces import ITALConditionable
+from eea.facetednavigation.interfaces import IFacetedNavigable
+from ..vocabulary import DashboardCollectionsVocabulary
 
 
 class TestConditionAwareVocabulary(IntegrationTestCase):
@@ -14,12 +17,14 @@ class TestConditionAwareVocabulary(IntegrationTestCase):
     def setUp(self):
         """Custom shared utility setup for tests."""
         self.portal = self.layer['portal']
+        self.folder = api.content.create(id='f', type='Folder', title='My category', container=self.portal)
         self.dashboardcollection = api.content.create(
             id='dc1',
             type='DashboardCollection',
             title='Dashboard collection 1',
-            container=self.portal
+            container=self.folder
         )
+        alsoProvides(self.folder, IFacetedNavigable)
 
     def test_conditionawarecollectionvocabulary(self):
         """This vocabulary is condition aware, it means
@@ -69,9 +74,19 @@ class TestConditionAwareVocabulary(IntegrationTestCase):
 
     def test_collectionsvocabulary(self):
         """This will return every DashboardCollections of the portal."""
+        # factory = queryUtility(IVocabularyFactory, u'imio.dashboard.collectionsvocabulary')
+        # one DashboardCollection
+        factory = DashboardCollectionsVocabulary()
+        self.assertEquals(len(factory(self.portal)), 1)
+        term = factory(self.portal).getTerm(self.dashboardcollection.UID())
+        self.assertEquals(term.token, term.value, self.dashboardcollection.UID())
+        self.assertEquals(term.title, self.dashboardcollection.Title())
+
+    def test_categorycollectionsvocabulary(self):
+        """This will return every DashboardCollections of the portal prefixed by categories."""
         factory = queryUtility(IVocabularyFactory, u'imio.dashboard.collectionsvocabulary')
         # one DashboardCollection
         self.assertEquals(len(factory(self.portal)), 1)
         term = factory(self.portal).getTerm(self.dashboardcollection.UID())
         self.assertEquals(term.token, term.value, self.dashboardcollection.UID())
-        self.assertEquals(term.title, self.dashboardcollection.Title())
+        self.assertEquals(term.title, '%s - %s' % (self.folder.Title(), self.dashboardcollection.Title()))